+++ stdlib/condition.mli
