+++ stdlib/mutex.mli
