+++ stdlib/semaphore.mli
