+++ stdlib/stdlib.ml
@@ -605,6 +605,7 @@ module BytesLabels  = BytesLabels
 module Callback     = Callback
 module Char         = Char
 module Complex      = Complex
+module Condition    = Condition
 module Digest       = Digest
 module Domain       = Domain
 module Either       = Either
@@ -626,6 +627,7 @@ module ListLabels   = ListLabels
 module Map          = Map
 module Marshal      = Marshal
 module MoreLabels   = MoreLabels
+module Mutex        = Mutex
 module Nativeint    = Nativeint
 module Obj          = Obj
 module Oo           = Oo
@@ -638,6 +640,7 @@ module Queue        = Queue
 module Random       = Random
 module Result       = Result
 module Scanf        = Scanf
+module Semaphore    = Semaphore
 module Seq          = Seq
 module Set          = Set
 module Stack        = Stack
